[feat] 각 투구별 시도 추가
