+++ src/main/java/bowling/domain/BowlingConstant.java
@@ -4,6 +4,5 @@ public class BowlingConstant {
     public static final int MAXIMUM_DOWNED_PIN_NUM = 10;
     public static final int MINIMUM_DOWNED_PIN_NUM = 0;
 
-    public static final int MINIMUM_TURN = 1;
-    public static final int MAXIMUM_TURN = 10;
+    public static final int ALL_PIN_DOWN = 10;
 }
